requirements.txt
asgiref==3.6.0
coverage==7.2.3
Django==4.2
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-decouple==3.8
sqlparse==0.4.3
tomli==2.0.1
